// frontend/src/api/generated/schema.d.ts
export interface components {
    schemas: {
        Applications: {
            readonly id: number;
            readonly owner: components["schemas"]["UserShort"];
            readonly center: components["schemas"]["CentersRetrieve"];
            readonly course: components["schemas"]["Courses"];
            content?: string | null;
            readonly index: number | null;
            center_id: number;
            course_id: number;
            /** Format: date-time */
            readonly created_date: string;
        };
        CentersList: {
            readonly id: number;
            name: string;
            slug?: string | null;
            info?: string | null;
            phone_number?: string | null;
            /** Format: uri */
            picture?: string | null;
            readonly detail_url: string;
        };
        CentersRetrieve: {
            readonly id: number;
            name: string;
            info?: string | null;
            phone_number?: string | null;
            phone_number_extra?: string | null;
            readonly courses: components["schemas"]["Courses"][];
            /** Format: int64 */
            cost?: number | null;
            /** Format: uri */
            official_website?: string | null;
            /** Format: uri */
            picture?: string | null;
            course_ids: number[];
        };
        Courses: {
            readonly id: number;
            title: string;
            slug?: string | null;
        };
        PatchedApplications: {
            readonly id?: number;
            readonly owner?: components["schemas"]["UserShort"];
            readonly center?: components["schemas"]["CentersRetrieve"];
            readonly course?: components["schemas"]["Courses"];
            content?: string | null;
            readonly index?: number | null;
            center_id?: number;
            course_id?: number;
            /** Format: date-time */
            readonly created_date?: string;
        };
        PatchedCentersRetrieve: {
            readonly id?: number;
            name?: string;
            info?: string | null;
            phone_number?: string | null;
            phone_number_extra?: string | null;
            readonly courses?: components["schemas"]["Courses"][];
            /** Format: int64 */
            cost?: number | null;
            /** Format: uri */
            official_website?: string | null;
            /** Format: uri */
            picture?: string | null;
            course_ids?: number[];
        };
        PatchedCourses: {
            readonly id?: number;
            title?: string;
            slug?: string | null;
        };
        TokenObtainPair: {
            username: string;
            password: string;
            readonly access: string;
            readonly refresh: string;
        };
        TokenRefresh: {
            readonly access: string;
            refresh: string;
        };
        UserShort: {
            /** @description Required. 150 characters or fewer. Letters, digits and @/./+/-/_ only. */
            username: string;
            first_name?: string;
            last_name?: string;
            phone_number: string;
            /**
             * Email address
             * Format: email
             */
            email?: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}